requirements_for_system.txt
Flask==2.0.1
mysql-connector-python==8.0.25
fpdf==1.7.2
schedule==1.1.0